requirements.txt
viam-sdk~=0.21.1
requests
pillow